Load a playlist when file dropped onto the main window
An adjustment was made to the chat pane's signals and slots. The
basic Print() is no longer what accepts chat messages. Instead,
this now has its own function which is attached explicitly to
the chat message receiver. Print() is now redirected to the status
section of the chat pane for messages coming from other subsystems.

Print()'s parameter was changed to const reference for efficiency
since const reference extend the life of the parameter anyway.