Fix profile picture size in EphemeralPanes
Utilize resizeEvent, but only if the widget isn't being destroyed
(which sets the size to 1x1). Add an internal flag that helps
determine if the Finished() signal has been called so we know when
we no longer need to resize the profile picture.

This technique has the added advantage of resizing the picture when
an OBS scene switch changes Celeste's dimensions through Pulsar when
that day comes.