Move numeric limits check in Random namespace
It doesn't make sense to check the int parameter against
std::numeric_limits because it has already been converted to an int
at this point, which means it has already overflowed if it's too big.
Check the size of the container, while it is still a std::size_type,
against std::numeric_limits.