Require parent objects for panes
There are memory leaks in the panes primarily because QMediaPlayer was
being instantiated without a parent. To prevent problems with the panes
in the future, require a pointer to a parent rather than defaulting to
nullptr.

Fixing this may have inadvertently fixed focus stealing on Windows
because creating a QWidget without a parent apparently causes the Qt
framework to create a new parent window for the widget.