Add pane accent color to Options UI
This require refactoring some things. Accent color was moved from
the Window class to the AnnouncePane class, which is the base class
for any panes that use the accent color. The main window doesn't
use an accent color.

Saving options was implemented. This required switching the source in
the BasicSettings class from a unique_ptr to a shared_ptr to make
BasicSettings copyable, otherwise the refrence to the settings in the
temporary AnnouncePane that has to be constructed to grab the settings
becomes invalid.