Refactor random selection of videos
Modified File::List class to provide its own random filename, when
applicable, from its internal list. Added support for filtering files
by extension when building the internal list. Item is selected from an
iterator to a shuffled to eliminated the potential of duplicates.

Added an option to commands.json and it's UI to allow duplicates or not.

Eliminated double lookup when forwarding command to
StageRedemptionCommand().