Delay loading options dialog until it's called for
Modify music player so it can be created without a parent, allowing
it to be created on the stack sanely.

Pass music player into bot class like what is done with security
object instead of creating it inside the bot class and trying
to fish it out for the dialog.