Lazy load commands dialog
This turned out to be quite a chore. Lazy loading the commands
dialog revealed a HUGE performance issue that HAD to be fixed.
A large number of commands resulted in 10+ second loading times,
which obviously would be unacceptable to a user. It was found this
was due to the large number of QWidgets that get created for each
entry.

Now, details for each command aren't loaded until the first time
they are expanded. Everything is shown collapsed by default.