Replace direct auth with Rewire server
A central server has been set up that will now hold the client
secret and handle obtaining and refreshing auth tokens. This server
is running an mqtt broker that will report the new token any time
a token is refreshed.

This should allow Celeste to authorize transparently most of the
time and not require any kind of server component in the bot
itself, making it easier for users to get started.

I'm utilizing the state query parameter as the session ID for
bot instances so Rewire can tell each user's bot apart.

On launch, tokens are retrieved from the server and validated
every hour, fulfilling a key Twitch requirement that I had been
overlooking. This further simplifies checking that everything is
ready to connect and makes Celeste launch more smoothly.

This is the first step toward replacing webhooks with websockets
in EventSub and eliminating the need for lighttpd in front of
Celeste.