Delay deleting profile image object
Call deleteLater() after emitting the signal
so there is something in the event loop that holds
off deleting the object until something is done with
the profile image in a slot.

It's hard to believe this bug has existed this long
but it's apparently from the colossal refactor
I did back in 2021.

Hopefully,
Closes #163