Multiple logger improvements
Fixes the following issues:
- Log messages from the logger itself were making
it to the UI, but weren't making it to the file.
- Writing directly to the logger by calling the
write slot from outside of the class breaks the
design. Instead, send errors messages as signals
from the class generating the error.
- Current log file was being overwritten on
subsequent runs on the same day. Clear current
and start over, appending to dated archive log
file.
- Rename was emptying the dated file when first
creating one because file position was being
moved back to the beginning when unnecessary to
do so.